Move sdp and usb, and fixups
